+++ 1.20.4/Common/src/main/java/fuzs/mindfuldarkness/client/gui/screens/PixelConfigScreen.java
@@ -102,6 +102,17 @@ public class PixelConfigScreen extends Screen {
     }
 
     @Override
+    public boolean keyPressed(int keyCode, int scanCode, int modifiers) {
+        if (super.keyPressed(keyCode, scanCode, modifiers)) {
+            return true;
+        }  else if (this.minecraft.options.keyInventory.matches(keyCode, scanCode)) {
+            this.onClose();
+            return true;
+        }
+        return false;
+    }
+
+    @Override
     public boolean isPauseScreen() {
         return false;
     }